Update it all
