+++ gradle.properties
@@ -6,7 +6,7 @@ org.gradle.parallel=true
 # Check these on https://fabricmc.net/develop/
 minecraft_version=1.20.1
 yarn_mappings=1.20.1+build.10
-loader_version=0.15.11
+loader_version=0.16.10
 
 # Mod Properties
 mod_version=1.6.3
@@ -14,4 +14,4 @@ maven_group=draylar
 archives_base_name=go-fish
 
 # Dependencies
-fabric_version=0.92.2+1.20.1
+fabric_version=0.92.3+1.20.1
